+++ __tests__/compute/index.ts
@@ -1,1 +1,6 @@
-export * as backendService from "./backendService"
+import * as gcp from "@pulumi/gcp";
+
+export const backendService = new gcp.compute.BackendService("fail#1", {
+    name: "fail#1",
+    loadBalancingScheme: "EXTERNAL"
+});