+++ src/cloudrunv2/index.ts
@@ -1,5 +1,8 @@
 import { disallowPublicIngress } from "./disallowPublicIngress";
+import { serviceDisallowEnvsSecrets, jobDisallowEnvsSecrets } from "./disallowEnvsSecrets";
 
 export const cloudrunv2Policies = [
+	serviceDisallowEnvsSecrets,
+	jobDisallowEnvsSecrets,
 	disallowPublicIngress,
 ];